Add `weekStartDate` attribute to partition date
This adds a new attribute to the items stored in dynamodb - `weekStartDate`. This is the date string (`yyyy-mm-dd`) on the *Monday* at the start of the current week.

Once we retrospectively add this to the existing entries in the db, we can create a new secondary index that uses this as its partition key. That will make it much easier to query the data to get entries for a given week.

There isn't a clean way to get 'the last 7 days worth', but with this approach if you calculate the last two Monday dates, you can use those to query the new index and get the last week + a bit of entries.