pg: Drop week identifying columns
Now we've switched to a database with a `DATE` type, it's rather easy to
query based on dates, and group them by them too. As such the column to
identify which week something was added is not needed and can be
dropped.

This change has a beneficial side effect of dropping mixed case column
names. PG doesn't like it, and it makes querying the data harder.