pg: Add primary key
While for storing the data there's little use for a primary key, the key
is needed to query the data. Using this primary key, composed of two
elements, we'd get a index for free on just `created_at` and it allows
for multiple pings to be added at the same time. _Except_ when it's the
same instance.