pg: Drop isDev column
Inferred data can be stored in a database, but in this instance I think
it's better not to. Now if there's a production install that
has been build with a git version, not a tag, it throws off our metrics.

It's easier to do this on the analytics side, and handle data
inconsistencies there. Another case this fixes for analytics; now the
bool can no longer be NULL. That's handled by another team.