collection: Don't SHA the Request IP
Right now the request IP is SHA'ed, which is an inadequate way of
protecting against FlowFuse knowing where the request came from. With a
small Go program I can iterate through IPs in about 4 minutes and build
a rainbow table.

As such, let's not spend the compute and just remove this.